@@ -2,25 +2,26 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using DualAuthCore.Models;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 
 namespace DualAuthCore.Data
 {
-    public class DualAuthContext : IdentityDbContext<ApplicationUser>
+  public class DualAuthContext : IdentityDbContext<ApplicationUser>
+  {
+    public DualAuthContext(DbContextOptions<DualAuthContext> options)
+        : base(options)
     {
-        public DualAuthContext(DbContextOptions<DualAuthContext> options)
-            : base(options)
-        {
-        }
 
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            base.OnModelCreating(builder);
-            // Customize the ASP.NET Identity model and override the defaults if needed.
-            // For example, you can rename the ASP.NET Identity table names and more.
-            // Add your customizations after calling base.OnModelCreating(builder);
-        }
     }
+
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+      base.OnModelCreating(builder);
+      // Customize the ASP.NET Identity model and override the defaults if needed.
+      // For example, you can rename the ASP.NET Identity table names and more.
+      // Add your customizations after calling base.OnModelCreating(builder);
+    }
+  }
 }
